+++ linuxdoom-1.10/ib_video.h
@@ -32,7 +32,7 @@ void IB_StartTic (void);
 void IB_GetFramebuffer(unsigned char **pixels, size_t *pitch);
 void IB_FinishUpdate (void);
 void IB_GetColor(unsigned char *bytes, unsigned char red, unsigned char green, unsigned char blue);
-void IB_InitGraphics(const char *title, size_t screen_width, size_t screen_height, unsigned int *bytes_per_pixel);
+void IB_InitGraphics(const char *title, size_t screen_width, size_t screen_height, size_t *bytes_per_pixel);
 void IB_ShutdownGraphics(void);
 void IB_GrabMouse(boolean grab);
